+++ src/components/BungaeCard.js
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+
 import styled, { css } from "styled-components";
 
 const StyledBungaeCard = styled.li`
