+++ src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
+import GlobalStyles from "./@styles/GlobalStyles";
 import App from "./App";
-import GlobalStyles from "./styles/GlobalStyles";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
