+++ src/pages/Root.js
@@ -1,14 +1,15 @@
-import { Outlet } from "react-router-dom";
+import { Outlet, useLocation } from "react-router-dom";
 
 import MainNavigation from "../components/MainNavigation";
 
 function RootLayout() {
+  const { pathname } = useLocation();
+  console.log(pathname);
+
   return (
     <>
       <MainNavigation />
-      <main>
-        <Outlet />
-      </main>
+      <Outlet />
     </>
   );
 }
